# treasure-planet/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Локальный сервер разработки"
        }
    ],
    "paths": {
...
        "/api/getstats": {
            "post": {
                "tags": [
                    "Levels"
                ],
                "summary": "Получить статистику уровня для игрока",
                "operationId": "4c4c007235d551a07833754d63f28e60",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id",
                                    "level_id"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "level_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Score"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка: Не получилось найти"
                    }
                }
            }
        },
        "/api/getlevels": {
            "post": {
                "tags": [
                    "Levels"
                ],
                "summary": "Получить список уровней игрока",
                "operationId": "87c8a6c74d19a93cfb32d9ef641002d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Список уровней игрока",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "levels": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Score"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка: Не получилось найти"
                    }
                }
            }
        },
        "/api/savelevelstats": {
            "post": {
                "tags": [
                    "Levels"
                ],
                "summary": "Сохранить статистику уровня для игрока",
                "operationId": "34c347d7519ebd1156b01f3f0294d046",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "player_id",
                                    "level_id",
                                    "score",
                                    "count_of_stars",
                                    "completed"
                                ],
                                "properties": {
                                    "player_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "level_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "score": {
                                        "type": "integer",
                                        "example": 1500
                                    },
                                    "count_of_stars": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "completed": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешно сохранено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Данные сохранены"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сохранения"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Score": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 5
                    },
                    "player_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "level_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "score": {
                        "type": "integer",
                        "example": 1500
                    },
                    "count_of_stars": {
                        "type": "integer",
                        "example": 3
                    },
                    "completed": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-17T12:34:56Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-17T12:34:56Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}